requirements.txt
Django==4.2.6
pre-commit==3.5.0
virtualenv==20.24.6
psycopg==3.1.12
psycopg-binary==3.1.12
django-htmx==1.17.0
django-crispy-forms==2.1
crispy-bootstrap4==2023.1
pytest-django==4.7.0
pytest-factoryboy==2.6.0
pytest-cov==4.1.0
djangorestframework==3.14.0
